+++ source/engine/component/InputComponent.h
@@ -15,7 +15,7 @@ public:
 
 	void registerInputCallback(int keyCode, void* function);
 	std::multimap<int, std::vector<std::function<void()>*>>& getKeyboardInputCallbackImpl();
-	std::multimap<int, std::vector<std::function<void(float)>*>>& getMouseInputCallbackImpl();
+	std::multimap<int, std::vector<std::function<void(double)>*>>& getMouseInputCallbackImpl();
 
 private:
 	enum moveDirection { FORWARD, BACKWARD, LEFT, RIGHT };
@@ -27,21 +27,21 @@ private:
 
 	void move(moveDirection moveDirection);
 
-	void rotateAroundXAxis(float offset);
-	void rotateAroundYAxis(float offset);
+	void rotateAroundXAxis(double offset);
+	void rotateAroundYAxis(double offset);
 
 	std::multimap<int, std::vector<std::function<void()>*>> m_keyboardInputCallbackImpl;
-	std::multimap<int, std::vector<std::function<void(float)>*>> m_mouseMovementCallbackImpl;
+	std::multimap<int, std::vector<std::function<void(double)>*>> m_mouseMovementCallbackImpl;
 
 	std::function<void()> f_moveForward;
 	std::function<void()> f_moveBackward;
 	std::function<void()> f_moveLeft;
 	std::function<void()> f_moveRight;
-	std::function<void(float)> f_rotateAroundXAxis;
-	std::function<void(float)> f_rotateAroundYAxis;
+	std::function<void(double)> f_rotateAroundXAxis;
+	std::function<void(double)> f_rotateAroundYAxis;
 
-	float moveSpeed = 0.05f;
-	float rotateSpeed = 2.0f;
+	double moveSpeed = 0.05f;
+	double rotateSpeed = 2.0f;
 
 };
